+++ cv_next/app/cv/[cvId]/page.tsx
@@ -2,16 +2,15 @@
 
 import { notFound } from "next/navigation";
 import { Metadata } from "next";
-import { getCvById, getRandomizedCvs } from "@/server/api/cvs";
+import { getCvById } from "@/server/api/cvs";
 import { decodeValue } from "@/lib/utils";
 import { ScrollToTop } from "@/components/ui/scrollToTop";
 import { CvPreview } from "@/components/ui/cvPreview";
 import { getCurrentId, userIsAdmin } from "@/server/api/users";
-import CarouselStructure from "@/components/ui/Carousel/carouselStructure";
-import { CVClickableCard } from "@/components/ui/CVComponents/cvClickableCard";
 import CommentsSection from "./components/commentSection/commentsSection";
 import CommentForm from "./components/commentSection/commentForm";
 import CvData from "./components/cvData";
+import { RecommendedCvsSection } from "./components/recommendedCvs";
 
 /**
  *
@@ -72,17 +71,6 @@ export default async function Page({ params }: { params: { cvId: string } }) {
   const authorData = JSON.parse(JSON.stringify(cv.user_id));
   const resultAdminCheck = await userIsAdmin();
 
-  const recommendedRaw = await getRandomizedCvs(true, 5, {
-    searchValue: "",
-    categoryIds: cv.cv_categories,
-  });
-  let recommendedCVs = (recommendedRaw || []).filter(
-    (rec: CvModel) => rec.id !== cv.id
-  );
-  if (recommendedCVs.length >= 5) {
-    recommendedCVs = recommendedCVs.slice(0, 5);
-  }
-
   return (
     <div>
       <ScrollToTop />
@@ -111,21 +99,7 @@ export default async function Page({ params }: { params: { cvId: string } }) {
         )}
       </div>
 
-      {recommendedCVs.length > 0 && (
-        <div className="bg-gradient mt-8 rounded-xl py-8">
-          <h2 className="mb-4 text-center text-3xl font-bold">Similar CVs</h2>
-          <h3 className="mb-4 text-center text-base font-bold text-gray-700 dark:text-gray-300">
-            {`Based on the categories of ${authorData.display_name}'s CV`}
-          </h3>
-          <CarouselStructure options={{ loop: true }}>
-            {recommendedCVs.map((recCv: CvModel) => (
-              <div className="p-4" key={recCv.id}>
-                <CVClickableCard cv={recCv} />
-              </div>
-            ))}
-          </CarouselStructure>
-        </div>
-      )}
+      <RecommendedCvsSection currentCv={cv} />
     </div>
   );
 }
